+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.53",
+  sdkVersion: "3.14.0-rc.54",
   genVersion: "2.730.0",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.53 2.730.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.54 2.730.0 2.0 @orq-ai/node",
 } as const;